requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-weasyprint==2.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.57.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
